Ugh. In older versions of OS X, array indexes are 32-bit. These days they're 64-bit on 64-bit systems and 32-bit on 32-bit systems. LLVM does a magic conversion between the two so old code still compiles.
This means that they almost always work, except that NSNotFound's value has also changed to a very large value on 64-bit systems. This converts to -1, but (int) NSNotFound != NSNotFound so anything that *uses* NSNotFound no longer works. The compiler is too magic to warn about this, so you just wind up with a broken program. Here I've fixed the most apparent issue this causes with Inform, but there are probably issues all over the place that can't be found due to the suppressed compiler warning.

(I had to fix this in Zoom too, and there's literally no way to tell the compiler to stop being so monumentally stupid and actually warn about the loss of accuracy. Grrrr. Setting the target scheme to 32-bits masks this issue too. Cunningly Apple have made this the default to make sure lots of broken software gets shipped)